SOAP-2770 Add functionality for uploading a trust store
Previously saved keystores are now defaulting to the Keystore crypto type for backwards compatibility.
Also changed the password dialog for key passwords to be a passwordPrompt, and thus not exposing the password.
Finally, did some minor clean up of commented out and unused code.

OBS! The GUI changes is only POC and needs to be reviewed.