Soapui tests call services on a local server
A new project is provided; it defines both mock and tests.
Tests now call the local server which exposes the mock services, thus we
do not rely on external services (the former was very instable) and the
test does not need any network access.

In addition, we now also check that the mock goal is working ;-)