Refactor SimpleForm and SimpleBindingForm to support adding multiple components in a row
The SimpleForm and SimpleFormBinding has become very messy so this is an initial
attempt to clean them up and also make them more flexible when it comes to defining
the form layout. The new functionality is now used by the OAUth 2Get Access Token Form.

SOAPUI-1640 Implement the Token status scenario