Do not pre-start (leak) mock service if test step is disabled
This affects mock services on the disabled tests refered by "Start Step"
from other WsdlMockResponseTestSteps. If mock service in the disabled step
is started, it might never be stopped as finish() is not called on
disabled steps. This results into mock service leak in the
JettyMockEngine which will wreack havoc when running other tests or
esp. repeating the same test.